* ext/extmk.rb.in (create_makefile): create def file only if   it does not yet exist. * lib/mkmf.rb: ditto.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/ruby_1_6@1304 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ lib/mkmf.rb
@@ -380,8 +380,10 @@ def create_makefile(target, srcdir = File.dirname($0))
 
   defflag = ''
   if RUBY_PLATFORM =~ /cygwin|mingw/
-    open(target + '.def', 'wb') do |f|
-      f.print "EXPORTS\n", "Init_", target, "\n"
+    if not File.exist? target + '.def'
+      open(target + '.def', 'wb') do |f|
+        f.print "EXPORTS\n", "Init_", target, "\n"
+      end
     end
     defflag = "--def=" + target + ".def"
   end
